resstrict

+++ notification/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect
 from notification.models import Notification
+from django.contrib import messages
 
 def notification_view(request):
     notifications = Notification.objects.all
@@ -7,12 +8,16 @@ def notification_view(request):
     return render(request, "notifications.html", context)
 
 def notification_delete(request,pk):
+    if request.user.is_superuser:
+        notifications = Notification.objects.get(id=pk)
 
-    notifications = Notification.objects.get(id=pk)
+        if request.method == 'POST':
+            notifications.delete()
+            return redirect('home')
+            
+        context = {'object': notifications}
+        return render(request, "delete_notification.html", context)
+    else:
+        messages.success(request, "You don't have permission")
+        return redirect("/")
 
-    if request.method == 'POST':
-        notifications.delete()
-        return redirect('home')
-        
-    context = {'object': notifications}
-    return render(request, "delete_notification.html", context)
